test(worker): split rooms route coverage from matchmaking
