+++ apps/worker/src/routes/matchmaking.test.mjs
@@ -7,7 +7,6 @@ import {
   handlePostMatchmakingQueue,
   matchMatchmakingQueueTicketPath,
 } from "./matchmaking.ts";
-import { handlePostRooms } from "./rooms.ts";
 
 function createEnv(doFetch) {
   return {
@@ -48,40 +47,6 @@ function createEnv(doFetch) {
     MIN_SUPPORTED_CLIENT_VERSION: "1.2.0",
     ISSUE_TOKEN: "",
     DISCORD_WEBHOOK_URL: "",
-  };
-}
-
-function createRoomsEnv() {
-  const initPayloads = [];
-  const lobbyUpserts = [];
-  const env = createEnv(async () => new Response("{}", { status: 200 }));
-
-  env.ROOM_DO.get = () => ({
-    fetch: async (request) => {
-      const url = new URL(request.url);
-      if (url.pathname === "/internal/init" && request.method === "POST") {
-        initPayloads.push(await request.json());
-        return new Response(JSON.stringify({ ok: true }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ error: "unexpected room request" }), { status: 400 });
-    },
-  });
-
-  env.LOBBY_DIRECTORY_DO.get = () => ({
-    fetch: async (request) => {
-      const url = new URL(request.url);
-      if (url.pathname === "/internal/upsert" && request.method === "POST") {
-        lobbyUpserts.push(await request.json());
-        return new Response(JSON.stringify({ ok: true }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ ok: true }), { status: 200 });
-    },
-  });
-
-  return {
-    env,
-    getInitPayloads: () => initPayloads,
-    getLobbyUpserts: () => lobbyUpserts,
   };
 }
 
@@ -213,54 +178,4 @@ test("handleGetMatchmakingWaitingCount rejects invalid query", async () => {
   assert.equal(called, false);
   const payload = await response.json();
   assert.equal(payload.error?.code, "BAD_REQUEST");
-});
-
-test("handlePostRooms does not upsert lobby for PUBLIC auto-match rooms", async () => {
-  const { env, getInitPayloads, getLobbyUpserts } = createRoomsEnv();
-  const response = await handlePostRooms(
-    new Request("https://example.com/api/rooms", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({
-        visibility: "PUBLIC",
-        auto_match: true,
-        mode: "ARENA",
-        win_metric: "SCORE",
-        play_style: "SP",
-        level_filter: "ANY",
-        max_players: 2,
-        room_comment: "Auto match room",
-      }),
-    }),
-    env,
-  );
-
-  assert.equal(response.status, 201);
-  assert.equal(getInitPayloads().length, 1);
-  assert.equal(getInitPayloads()[0]?.settings.auto_match, true);
-  assert.equal(getLobbyUpserts().length, 0);
-});
-
-test("handlePostRooms keeps lobby upsert for normal PUBLIC rooms", async () => {
-  const { env, getLobbyUpserts } = createRoomsEnv();
-  const response = await handlePostRooms(
-    new Request("https://example.com/api/rooms", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({
-        visibility: "PUBLIC",
-        mode: "ARENA",
-        win_metric: "SCORE",
-        play_style: "SP",
-        level_filter: "ANY",
-        max_players: 2,
-        room_comment: "Public room",
-      }),
-    }),
-    env,
-  );
-
-  assert.equal(response.status, 201);
-  assert.equal(getLobbyUpserts().length, 1);
-  assert.equal(getLobbyUpserts()[0]?.isPublic, true);
 });
